pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.5"
python-dotenv = "^0.14.0"
requests = "^2.31.0"
pytest = "^8.1.1"
selenium = "^4.18.1"
pymongo = "^4.10.1"
mongomock = "^4.2.0.post1"
safety = "^3.2.10"
loggly-python-handler = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
